chatbot/backend/packages/simplechat/pyproject.toml
[tool.poetry]
name = "simplechat"
version = "0.0.1"
description = "Simple chat"
authors = ["Juan Calvo <juan.calvo@datatonic.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
langchain = ">=0.0.333"
google-cloud-aiplatform = ">=1.35.0"
google-cloud-discoveryengine = "^0.11.3"
wikipedia = "^1.4.0"
twilio = "^8.10.3"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"
fastapi = "^0.104.0"
sse-starlette = "^1.6.5"

[tool.langserve]
export_module = "rag_google_cloud_vertexai_search"
export_attr = "chain"

[tool.templates-hub]
use-case = "rag"
author = "Datatonic"
integrations = [ "Google"]
tags = ["vectordbs"]

[build-system]
requires = [
    "poetry-core",
]
build-backend = "poetry.core.masonry.api"
